# src/docs/openapi.yaml
openapi: '3.0.3'
info:
  version: 0.0.1
  title: Todo Mock API
servers:
  - url: /api
    description: Mock Server
security:
  - Bearer: []
paths:
  /users/register:
    post:
      tags:
        - users
      summary: ユーザーを登録します
      description: ユーザー名とパスワードを送信してユーザーを登録します
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: ユーザーの登録に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '409':
          description: 同じ名前のユーザーが既に存在します
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictUserError'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
  /users/login:
    post:
      tags:
        - users
      summary: ログイン
      description: ユーザー名とパスワードを送信してトークンを発行します
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: ログインに成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/post_users_login_response'
        '400':
          description: 不正なユーザーです
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidUserError'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MismatchedPasswordError'
        '404':
          description: ユーザーが見つかりませんでした
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNotFoundError'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
  /users/logout:
    post:
      tags:
        - users
      summary: ログアウト
      responses:
        '200':
          description: ログアウトに成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '401':
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        '400':
          $ref: '#/components/responses/TokenInvalidResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
...
components:
  schemas:
    token:
      description: 認証用トークン. token68
      type: string
      pattern: '/^[0-9a-zA-Z-._~+/]+=*$/'
...
    user:
      description: ユーザー登録に必要なデータ
      type: object
      properties:
        username:
          description: 氏名
          example: guest
          type: string
        password:
          description: パスワード
          example: password
          type: string
...
    success:
      description: 成功
      type: object
      properties:
        success:
          description: 成功フラグ
          example: true
          type: boolean
    post_users_login_response:
      description: 'ログイン成功時のレスポンス'
      type: object
      properties:
        success:
          description: 成功フラグ
          example: true
          type: boolean
        token:
          $ref: '#/components/schemas/token'
    Error:
      description: エラー
      type: object
      properties:
        code:
          description: エラーコード
          type: string
        message:
          description: エラーメッセージ
          type: string
    ConflictUserError:
      description: 既に同じユーザーが存在します
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'ConflictUser'
        message:
          description: エラーメッセージ
          type: string
    MismatchedPasswordError:
      description: パスワードが一致しませんでした
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'MismatchedPassword'
        message:
          description: エラーメッセージ
          type: string
    InvalidUserError:
      description: ユーザーの値が不正です
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'InvalidUser'
        message:
          description: エラーメッセージ
          type: string
    UserNotFoundError:
      description: 対象のユーザーが見つかりませんでした
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'UserNotFound'
        message:
          description: エラーメッセージ
          type: string
    InvalidTokenError:
      description: トークンの値が不正です
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'InvalidToken'
        message:
          description: エラーメッセージ
          type: string
    TokenRequiredError:
      description: トークンが見つかりませんでした
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'TokenRequired'
        message:
          description: エラーメッセージ
          type: string
    MismatchedTokenError:
      description: トークンが一致しませんでした
      type: object
      properties:
        code:
          description: エラーコード
          type: string
          enum:
            - 'MismatchedToken'
        message:
          description: エラーメッセージ
          type: string
...
  responses:
    DefaultErrorResponse:
      description: 操作に失敗しました
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    DefaultUnauthorizedErrorResponse:
      description: 認証に失敗しました
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/MismatchedTokenError'
              - $ref: '#/components/schemas/TokenRequiredError'
...
    TokenInvalidResponse:
      description: トークンの値が不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InvalidTokenError'
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: API へのアクセスに必要なトークン。結構適当です。
      bearerFormat: token68
...